fix(ecosystems): tolerate commit_stats.dds as a JSON string
ecosyste.ms serializes dds as a quoted string for some repos and a number for
others; the float64 field rejected the string form and the repo was skipped.
Parse via a flexFloat that accepts both.

+++ internal/db/repos.go
@@ -156,7 +156,7 @@ func (d *DB) SetEnrichment(fullName string, r *ecosystems.Repository) error {
 		if r.CommitStats != nil {
 			commits = int64(r.CommitStats.TotalCommits)
 			comtrs = int64(r.CommitStats.TotalCommitters)
-			dds = r.CommitStats.DDS
+			dds = float64(r.CommitStats.DDS)
 		}
 		if score, ok := r.ScorecardScore(); ok {
 			scorecardVal = score
